% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedblockchain_operations.R
\name{managedblockchain_vote_on_proposal}
\alias{managedblockchain_vote_on_proposal}
\title{Casts a vote for a specified ProposalId on behalf of a member}
\usage{
managedblockchain_vote_on_proposal(NetworkId, ProposalId, VoterMemberId,
  Vote)
}
\arguments{
\item{NetworkId}{[required] The unique identifier of the network.}

\item{ProposalId}{[required] The unique identifier of the proposal.}

\item{VoterMemberId}{[required] The unique identifier of the member casting the vote.}

\item{Vote}{[required] The value of the vote.}
}
\description{
Casts a vote for a specified \code{ProposalId} on behalf of a member. The
member to vote as, specified by \code{VoterMemberId}, must be in the same AWS
account as the principal that calls the action.
}
\details{
Applies only to Hyperledger Fabric.
}
\section{Request syntax}{
\preformatted{svc$vote_on_proposal(
  NetworkId = "string",
  ProposalId = "string",
  VoterMemberId = "string",
  Vote = "YES"|"NO"
)
}
}

\keyword{internal}
